add all

+++ app/Http/Controllers/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Auth;
 use App\Models\Task;
 
 class TaskController extends Controller
@@ -15,14 +16,23 @@ class TaskController extends Controller
      */
     public function index()
     {
-        //タスクの一覧を取得
-        $tasks = Task::all();
+        $data=[];
         
-        //タスク一覧ビューでそれを表示
-        return view("tasks.index",[
-            "tasks" => $tasks,
-        ]);
+        if (\Auth::check()) {    //認証済みの場合
+            $user = \Auth::user();
+            $status = $user->tasks()->orderBy("created_at","desc")->paginate(10);
+            $tasks = $user->tasks()->orderBy("created_at","desc")->paginate(10);
+            $data = [
+                "user" => $user,
+                "status" => $status,
+                "tasks" => $tasks,
+            ];
+        }
+        
+        //dashboardビューでそれら(タスク一覧)を表示
+        return view("dashboard", $data);
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -53,11 +63,11 @@ class TaskController extends Controller
             "status" => "required|max:10",
         ]);
         
-        //タスクを作成
-        $task = new Task;
-        $task ->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //認証済みユーザとしてタスクを作成
+        $request->user()->tasks()->create([
+            "status" => $request->status ,
+            "content" => $request->content ,
+        ]);
         
         //トップページへリダイレクトさせる
         return redirect("/");
@@ -115,7 +125,7 @@ class TaskController extends Controller
         $task->content = $request->content;
         $task->save();
         
-        return redirect("/");
+        return redirect('/');
     }
 
     /**
@@ -127,8 +137,15 @@ class TaskController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
         
-        return redirect("/");
+        //認証済みユーザがそのタスク登録者である場合はタスクを削除
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/')
+                ->with("success","Delete Successful");
+        }
+        
+        return redirect('/')
+            ->with("Delete Failed");
     }
 }
